vfs: improve setup routine, move it into a separate file
