+++ src/vfs/setup.h
@@ -1,0 +1,9 @@
+
+#ifndef TAISEI_VFS_SETUP
+#define TAISEI_VFS_SETUP
+
+#include "util.h"
+
+void vfs_setup(bool silent);
+
+#endif
